Work for https://github.com/surveyjs/survey-creator/issues/1680 - implemented createResponsivityManager property
