+++ src/actions/container.ts
@@ -1,6 +1,6 @@
 import { propertyArray } from "../jsonobject";
 import { Base } from "../base";
-import { Action } from "./action";
+import { IAction, Action } from "./action";
 
 export class ActionContainer<T extends Action = Action> extends Base {
     @propertyArray({
@@ -52,5 +52,22 @@ export class ActionContainer<T extends Action = Action> extends Base {
     }
     public get css(): string {
       return "sv-action-bar" + (!!this.containerCss ? " " + this.containerCss : "");
+    }
+
+    private sortItems(items: Array<T>) {
+      return []
+        .concat(items.filter((item) => item.visibleIndex >= 0 || item.visibleIndex === undefined))
+        .sort((firstItem, secondItem) => {
+          return firstItem.visibleIndex - secondItem.visibleIndex;
+        });
+    }
+
+    public setItems(items: Array<IAction>, sortByVisibleIndex = true): void {
+      const actions: Array<T> = <any>items.map((item) => (item instanceof Action ? item : new Action(item)));
+      if (sortByVisibleIndex) {
+        this.actions = this.sortItems(actions);
+      } else {
+        this.actions = actions;
+      }
     }
 }
